refactor: remove deprecated openai-error-handler and consolidate into error-handler
Replace all imports of the deprecated `openai-error-handler` shim with
`error-handler`, which has been the canonical source since #10204.
Delete `openai-error-handler.ts` and its duplicate test file, as
`error-handler.spec.ts` already covers the same behavior.
